Bases are built somewhat properly
Bases dont build inside the main base anymore. Tries to find the nearest  resources in a certain range. Added some micro managing for how many units are to be made given the amount of command centers built. This ensures that an economy can be built up for another command center (costs 400 to make)